+++ NEWS.md
@@ -1,5 +1,7 @@
 # sds dev
 
+* New source `usgs_seamless()` elevation from the US. 
+
 * `CGAZ_sql()` now works with no arguments, and returns a general query useable by vapour_read_geometry() and friends. 
 
 * Add `rema()` and `rema_v2()`.